test reshuffled
