+++ tests/core/test_registry.py
