+++ tests/diagnostics/test_bands.py
