+++ tests/diagnostics/test_common.py
