+++ tests/diagnostics/test_registry.py
