+++ tests/diagnostics/test_trend_core.py
