+++ tests/display/test_renderer_contracts.py
