+++ tests/display/test_step_memory_formatter.py
