+++ tests/reporting/summary/test_fixtures.py
