+++ tests/telemetry/test_msgpack.py
