+++ tests/telemetry/test_sender_sequence.py
